# [chore] Reuse the JSON writer to reduce memory allocations