@atoulme 2023-10-19
Closing ; we can revisit this whole logic once exporterhelper is there.